Text input does not work this way.
reading keypresses is not enough to determine text.
Char.fromCode gives me capital letters, and no value
for things like space or dash.

Looking at todomvc, I need to do something like this
to get the -browser- to interpret the typing in the box
and supply me with a string.

on "input" targetValue (Signal.message address << UpdateField)

So now I'm gonna be on to addresses after all.